ml/smolvlm-training/requirements.txt
transformers
qwen-vl-utils
torch
torchvision
accelerate
peft
python-dotenv
trl
bitsandbytes==0.43.0
num2words